@@ -2,16 +2,16 @@
   "name": "hksanda-website",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
     "lint": "vue-cli-service lint",
-    "api": "nodemon ./api/server.ts",
+    "api": "nodemon --watch api -e js,ts --loader=ts-node/esm api/server.js",
     "dev": "vite --host"
   },
   "dependencies": {
     "@apollo/client": "^3.5.6",
-    "@fortawesome/fontawesome-free": "^5.15.4",
     "@jamescoyle/vue-icon": "^0.1.2",
     "@mdi/js": "^6.5.95",
     "@vue/apollo-option": "^4.0.0-alpha.16",
@@ -28,14 +28,13 @@
     "graphql-upload": "^13.0.0",
     "gsap": "^3.9.1",
     "nanoid": "^3.1.30",
-    "nodemon": "^2.0.15",
     "opencc-js": "^1.0.3",
-    "primeflex": "2.0.0",
+    "primeflex": "3.1.2",
     "primeicons": "^5.0.0",
     "primevue": "3.10.0",
     "scrollmagic": "^2.0.8",
     "scrollmagic-plugin-gsap": "^1.0.4",
-    "swiper": "6",
+    "swiper": "^6",
     "vue": "^3.0.0",
     "vue-apollo": "^3.1.0",
     "vue-fullpage.js": "^0.1.7",
@@ -47,8 +46,14 @@
     "vue3-youtube": "^0.1.9"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^4.18.0",
-    "@typescript-eslint/parser": "^4.18.0",
+    "@babel/core": "^7.16.7",
+    "@babel/preset-env": "^7.16.8",
+    "@types/cors": "^2.8.12",
+    "@types/express": "^4.17.13",
+    "@types/graphql-upload": "^8.0.9",
+    "@types/node": "^17.0.8",
+    "@typescript-eslint/eslint-plugin": "5.9.1",
+    "@typescript-eslint/parser": "5.9.1",
     "@vitejs/plugin-react-refresh": "^1.3.6",
     "@vitejs/plugin-vue": "^2.0.1",
     "@vue/cli-plugin-babel": "~4.5.0",
@@ -56,20 +61,22 @@
     "@vue/cli-plugin-typescript": "~4.5.0",
     "@vue/cli-service": "~4.5.0",
     "@vue/compiler-sfc": "^3.0.0",
-    "@vue/eslint-config-typescript": "^7.0.0",
+    "@vue/eslint-config-typescript": "10.0.0",
     "apollo-upload-client": "^17.0.0",
     "babel-eslint": "^10.1.0",
-    "eslint": "^6.7.2",
+    "eslint": "8.6.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-vue": "^7.0.0",
+    "eslint-plugin-vue": "8.2.0",
     "graphql": "^16.2.0",
     "mongoose": "^6.1.4",
     "node-sass": "^6.0.1",
     "prettier": "^2.5.1",
     "sass": "^1.41.1",
     "sass-loader": "^10",
-    "typescript": "~4.1.5",
+    "ts-node": "^10.4.0",
+    "ts-node-dev": "^1.1.8",
+    "typescript": "4.5.4",
     "unplugin-auto-import": "^0.5.9",
     "unplugin-vue-components": "^0.17.11",
     "vite": "^2.7.10",
