@@ -1,11 +1,21 @@
-const Class = require("../models/class");
+import Class from "../models/class.js";
 
-exports.getClasses = async () => {
+interface ClassInput {
+  title: string;
+  type: string;
+  classroom?: string;
+  time: string;
+  students: string;
+  location?: string;
+  mapQuery?: string;
+}
+
+export const getClasses = async () => {
   return await Class.find();
 };
 
-exports.addClass = (_, args) => {
-  let details = new Class({
+export const addClass = (_: any, args: { details: ClassInput }) => {
+  const details = new Class({
     ...args.details,
   });
   return details.save();
