@@ -1,11 +1,9 @@
 module.exports = {
-  client: {
-    service: {
-      name: "hksanda-website-4ya9pk",
-      // URL to the GraphQL API
-      url: "http://localhost:8000/graphql",
-    },
-    // Files processed by the extension
-    includes: ["src/**/*.vue", "src/**/*.js"],
+  service: {
+    name: "hksanda-website-4ya9pk",
+    // URL to the GraphQL API
+    url: "http://localhost:8000/graphql",
   },
+  // Files processed by the extension
+  includes: ["src/**/*.vue", "src/**/*.js", "src/**/*.ts"],
 };
